rags/rag_agent_openai/requirements.txt
chromadb>=1.0.0
sentence-transformers>=4.0.0
langchain>=0.3.0
openai>=1.0.0
gradio>=5.0.0
fi-instrumentation
futureagi
traceAI-openai
opentelemetry>=2.0.0
python-dotenv>=1.0.0
